Multi-LLM Refinement Pipeline for Query Output Validation (#118)
* Multi-LLM Refinement Pipeline for Query Output Validation  (#92)

## Summary of PR

This PR contains the work to add a new `CriticConvert` physical operator to PZ. At a high-level, this operator runs a bonded convert, and then asks a critic model if the answer produced by the bonded convert can be improved upon. The original output and the critique are then fed into a refinement model, which produces the improved output.

The work to implement this includes:
1. Defining the physical operator in `src/palimpzest/query/operators/critique_and_refine_convert.py`
2. Adding an implementation rule for this physical operator in `src/palimpzest/query/optimizer/rules.py`
3. Adding boolean flag(s) to enable allowing / disallowing this physical optimization
4. Adding base prompts for the critique and refinement generations

One other change which this work spawned was an attempt to improve the management and construction of our prompts -- and to decouple this logic from the `BaseGenerator` class. On the management side, I split our single `prompts.py` file into a set of files. On the construction side, I created a `PromptFactory` class which templates prompts based on the `prompt_strategy` and input record.

The `PromptFactory` is not a perfect solution, but I think it is a step in the right direction.

Finally, I fixed an error which previously filtered out `RAGConvert` operators from being considered by the `Optimizer`, and I made 2-3 more miscellaneous small tweaks.

---------

Co-authored-by: Yash Agarwal <yash94404@gmail.com>
Co-authored-by: Yash Agarwal <yashaga@Yashs-Air.attlocal.net>