replace DataReader  with  ValidationData
ValidationData class can read in the execution_stats and produce what we need for the inputs and expected_data.

If users use annotation tools, ValidationData will read "annotation" field
If users don't want to do anything for the annotation, they just use execution_stats that was produced from previous round.

Note that:
Given that the currently the code doesn't really use expected_output, most of the time it just uses champion model's output, I suggest:
1. don't force val_data not None for mab/random sampling processors;
2. Use sample=n for dataset source instead of `dataset._set_data_source(self.val_data.input_dataset())`

THIS IS NOT READY TO SUBMIT.